pyproject.toml
[tool.poetry]
name = "neuro-posts"
version = "1.0"
description = "Telegram bot that uses a neural network to write posts to the channel"
authors = ["Dellenoam"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.34.0"
google-generativeai = "^0.4.1"
pydantic-settings = "^2.2.1"
aiofiles = "^23.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
